Fixes #10 and #11 - Working C-64 version with Joystick control! Many thanks to the cc65-devel mailing list for help tracking down the weird behavior I was having with the C-64 version.  Learned a lot with that.
For now, the C-64 version is played with a joystick in port 2, and the
X-16 version is still played with a mouse.

The only 'feature' missing on the C-64 version is the re-defined chars
used as an intermediary half-way on the piece rotation, need to poke
around the memory map to see if I have free space to do redefined chars
on the C-64 version still.

It should, theoretically, compile for the C-128 as well.